+++ src/renderer/src/App.tsx
@@ -1,6 +1,5 @@
 import { AppShell } from './components/AppShell'
 import { PromptProvider } from './components/PromptModal'
-import { BugReportProvider } from './components/BugReportModal'
 import { UpdateNotifier } from './components/UpdateNotifier'
 import { WorkspaceProvider } from './store/workspace'
 import { DiagnosticsProvider } from './store/diagnostics'
@@ -16,9 +15,7 @@ function App(): JSX.Element {
           <SyncProvider>
             <DiagnosticsProvider>
               <ConsoleProvider>
-                <BugReportProvider>
-                  <AppShell />
-                </BugReportProvider>
+                <AppShell />
                 <UpdateNotifier />
               </ConsoleProvider>
             </DiagnosticsProvider>